New project/backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.27.1
pydantic==2.6.1
pydantic-settings==2.2.1
SQLAlchemy==2.0.27
psycopg2-binary==2.9.9
alembic==1.13.1
python-jose[cryptography]==3.3.0
python-multipart==0.0.9
pandas==2.2.1
numpy==1.26.4
scikit-learn==1.4.1.post1
sentence-transformers==2.4.0
httpx==0.27.0
pytest==8.0.2
pytest-asyncio==0.23.5
